Refactor sequence into separate package.
First preparations for the separation of PersistenceLite and Sequence.
