+++ src/main/java/grevend/persistencelite/dao/BaseDao.java
@@ -27,8 +27,8 @@ package grevend.persistencelite.dao;
 import grevend.persistencelite.entity.EntityFactory;
 import grevend.persistencelite.entity.EntityMetadata;
 import grevend.persistencelite.util.ExceptionEscapeHatch;
-import grevend.persistencelite.util.function.ThrowingConsumer;
-import grevend.persistencelite.util.function.ThrowingFunction;
+import grevend.sequence.function.ThrowingConsumer;
+import grevend.sequence.function.ThrowingFunction;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
