+++ src/main/java/grevend/sequence/function/ThrowingBiConsumer.java
@@ -22,45 +22,13 @@
  * SOFTWARE.
  */
 
-package grevend.persistencelite.util.iterators;
+package grevend.sequence.function;
 
-import java.util.ArrayDeque;
-import java.util.Collection;
-import java.util.Iterator;
-import java.util.Queue;
-import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
-public class MergeIterator<T> extends ChainIterator<T> {
+@FunctionalInterface
+public interface ThrowingBiConsumer<T, U> {
 
-    private final Queue<Iterator<? extends T>> queue;
-
-    public MergeIterator(@NotNull Iterator<T> iterator, @NotNull Collection<Iterator<T>> iterators) {
-        super(iterator);
-        this.queue = new ArrayDeque<>();
-        this.queue.add(iterator);
-        this.queue.addAll(iterators);
-    }
-
-    @Override
-    public boolean hasNext() {
-        while (!this.queue.isEmpty()) {
-            if (this.queue.peek().hasNext()) {
-                return true;
-            }
-            this.queue.poll();
-        }
-        return false;
-    }
-
-    @Override
-    public T next() {
-        T element = null;
-        var iterator = this.queue.poll();
-        if (iterator != null) {
-            element = iterator.next();
-            this.queue.offer(iterator);
-        }
-        return element;
-    }
+    void accept(@Nullable T t, @Nullable U u) throws Exception;
 
 }
