+++ src/main/java/grevend/sequence/function/TriFunction.java
@@ -22,13 +22,13 @@
  * SOFTWARE.
  */
 
-package grevend.persistencelite.util.function;
+package grevend.sequence.function;
 
 import org.jetbrains.annotations.Nullable;
 
 @FunctionalInterface
-public interface ThrowingConsumer<T> {
+public interface TriFunction<T, U, S, R> {
 
-    void accept(@Nullable T t) throws Exception;
+    @Nullable R apply(@Nullable T t, @Nullable U u, @Nullable S s);
 
 }
